+++ rqt_gui_cpp/include/rqt_gui_cpp/plugin.hpp
