+++ rqt_gui_cpp/src/rqt_gui_cpp/roscpp_plugin_provider.hpp
